# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartHome IoT API",
        "description": "API documentation for SmartHome IoT system",
        "contact": {
            "email": "leducanh1503.works@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/device-history": {
            "get": {
                "tags": [
                    "Device History"
                ],
                "summary": "Get device toggle history",
                "description": "Retrieve the history of device toggles with optional search by time and pagination",
                "operationId": "getDeviceToggleHistory",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search_time",
                        "in": "query",
                        "description": "Search by time (accepts formats: DD/MM/YYYY, YYYY-MM-DD, HH:MM:SS, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "01/10/2024"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deviceHistory": {
                                            "description": "List of device toggle records",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "toggled_at": {
                                                        "type": "string",
                                                        "example": "2023-12-12 15:30:00"
                                                    },
                                                    "device_name": {
                                                        "type": "string",
                                                        "example": "Light Bulb"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "description": "Pagination details",
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}